Added a ZStack as an ImageCollection
ImageCollection is general, but ZStack is a particular ImageCollection: all images must be the same shape, and the represent a stacking in z of these images.

ImageCollection now sorts the file names before adding files so they can be added in a meaningful way (os.dir does not necessarily sorts everything alphabetically)

Image() now has a shape @property and it keeps the path as an accessible variable (self.path, not self.__path)

A function removeChannels() has been added to strip unwanted channels (often useful for analysis).