started some tests for ZStack... + Issue/comments:
It still feels weird to be handling a collection of Image objects while I have a 3D array ready that could be enough on its own. Since I decided to keep ZStack as a child of ImageCollection, i'm cutting the array and loading all layers as 2D image objects... Ok but it looks unecessary and probably time-consuming.

Would be nice to test ImageCollection all the way. I see there's a DCCImageCollection test file but it looks deprecated.