Modified database API
The database connects upon creation.

There is no need to have cryptic 'ro'  'rw' 'rwc' etc... We always want to read the database and always want to create if we have write permission and it does not exist.  An simple flag "writePermission=False" is added to __init__ and mode has been dropped. By default, it is readonly.

The changeConnectionMode function was replaced by setPermissionToReadOnly and setPermissionToWrite().  changeConnectionMode() was left for compatibility reasons.

begin() and end() were changed to beginTransaction() and endTransaction() for clarity.

All tests were modified accordingly.