Wixed incoherence with the matrix transposition and worked on angular average
There was some incoherence with the transposition: since most of the libraries we use work with array (Y, X), we need to transpose the pixel matrix before using these libraries. For example, skimage's sobel uses arrays, so we need to transpose self.pixels in order to get the vertical sobel instead of the horizontal sobel (and vice versa), i.e. we were calling sobel_h and the result was sobel_v...