Dropped support for windows path in PathPattern
We cannot easily support Windows path (with backslahes \\).
Normalize your path with "/" instead of "\" as a separator.
Windows path cannot be used directly: it is not possible to decide
the meaning of escaped characters:  C:\newbie\dccote\test.tiff  Is it a newline character?  Is it a \d for the regexp ? The solution could be to have the user escape everything but then the newline must be "unescaped" for the regex. This would be complicated.  For now, just drop support.