Update database.py
Found the bug in mysql.connector: for some reason, there a python and a C version that can be used.  By default, the C version will be used if available.  Connecting with use_pure=True will use a python-only version, and this one is fine, even with the latest from mysql.connector.python