Compelte rework of URL scheme
Instead of trying to come up with intelligent analysis of the url and create a tunnel when needed, I decided to do like ValentinaDB and others and use a different scheme: mysql or mysql+ssh. This required a bit a internal rework to clear up sshHost sshUser mysql host and mysql user.